backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
Flask-Migrate==4.0.7
alembic==1.13.2
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-WTF==1.2.1
python-dotenv==1.0.1
Flask-SocketIO==5.3.6
eventlet==0.33.3
celery==5.4.0
redis==5.0.8
pytest==8.3.3
Flask-Testing==0.8.1
Flask-Caching==2.3.0
